make SSH Agent support internal
