+++ lib/ansible/cli/__init__.py
@@ -9,8 +9,6 @@ import locale
 import os
 import sys
 
-from ansible._internal._errors import _alarm_timeout
-
 # We overload the ``ansible`` adhoc command to provide the functionality for
 # ``SSH_ASKPASS``. This code is here, and not in ``adhoc.py`` to bypass
 # unnecessary code. The program provided to ``SSH_ASKPASS`` can only be invoked
@@ -76,8 +74,6 @@ def initialize_locale():
 
 initialize_locale()
 
-
-import atexit
 import errno
 import getpass
 import subprocess
@@ -113,92 +109,23 @@ from ansible.module_utils.six import string_types
 from ansible.module_utils.common.text.converters import to_bytes, to_text
 from ansible.module_utils.common.collections import is_sequence
 from ansible.module_utils.common.file import is_executable
-from ansible.module_utils.common.process import get_bin_path
 from ansible.parsing.dataloader import DataLoader
 from ansible.parsing.vault import PromptVaultSecret, get_file_vault_secret, VaultSecretsContext
 from ansible.plugins.loader import add_all_plugin_dirs, init_plugin_loader
 from ansible.release import __version__
-from ansible.utils._ssh_agent import SshAgentClient
 from ansible.utils.collection_loader import AnsibleCollectionConfig
 from ansible.utils.collection_loader._collection_finder import _get_collection_name_from_path
 from ansible.utils.path import unfrackpath
 from ansible.vars.manager import VariableManager
 from ansible.module_utils._internal import _deprecator
+from ansible._internal._ssh import _agent_launch
+
 
 try:
     import argcomplete
     HAS_ARGCOMPLETE = True
 except ImportError:
     HAS_ARGCOMPLETE = False
-
-
-_SSH_AGENT_STDOUT_READ_TIMEOUT = 5  # seconds
-
-
-def _launch_ssh_agent() -> None:
-    ssh_agent_cfg = C.config.get_config_value('SSH_AGENT')
-    match ssh_agent_cfg:
-        case 'none':
-            display.debug('SSH_AGENT set to none')
-            return
-        case 'auto':
-            try:
-                ssh_agent_bin = get_bin_path(C.config.get_config_value('SSH_AGENT_EXECUTABLE'))
-            except ValueError as e:
-                raise AnsibleError('SSH_AGENT set to auto, but cannot find ssh-agent binary.') from e
-            ssh_agent_dir = os.path.join(C.DEFAULT_LOCAL_TMP, 'ssh_agent')
-            os.mkdir(ssh_agent_dir, 0o700)
-            sock = os.path.join(ssh_agent_dir, 'agent.sock')
-            display.vvv('SSH_AGENT: starting...')
-            try:
-                p = subprocess.Popen(
-                    [ssh_agent_bin, '-D', '-s', '-a', sock],
-                    stdin=subprocess.PIPE,
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.PIPE,
-                    text=True,
-                )
-            except OSError as e:
-                raise AnsibleError('Could not start ssh-agent.') from e
-
-            atexit.register(p.terminate)
-
-            help_text = f'The ssh-agent {ssh_agent_bin!r} might be an incompatible agent.'
-            expected_stdout = 'SSH_AUTH_SOCK'
-
-            try:
-                with _alarm_timeout.AnsibleTimeoutError.alarm_timeout(_SSH_AGENT_STDOUT_READ_TIMEOUT):
-                    stdout = p.stdout.read(len(expected_stdout))
-            except _alarm_timeout.AnsibleTimeoutError as e:
-                display.error_as_warning(
-                    msg=f'Timed out waiting for expected stdout {expected_stdout!r} from ssh-agent.',
-                    exception=e,
-                    help_text=help_text,
-                )
-            else:
-                if stdout != expected_stdout:
-                    display.warning(
-                        msg=f'The ssh-agent output {stdout!r} did not match expected {expected_stdout!r}.',
-                        help_text=help_text,
-                    )
-
-            if p.poll() is not None:
-                raise AnsibleError(
-                    message='The ssh-agent terminated prematurely.',
-                    help_text=f'{help_text}\n\nReturn Code: {p.returncode}\nStandard Error:\n{p.stderr.read()}',
-                )
-
-            display.vvv(f'SSH_AGENT: ssh-agent[{p.pid}] started and bound to {sock}')
-        case _:
-            sock = ssh_agent_cfg
-
-    try:
-        with SshAgentClient(sock) as client:
-            client.list()
-    except Exception as e:
-        raise AnsibleError(f'Could not communicate with ssh-agent using auth sock {sock!r}.') from e
-
-    os.environ['SSH_AUTH_SOCK'] = os.environ['ANSIBLE_SSH_AGENT'] = sock
 
 
 class CLI(ABC):
@@ -635,10 +562,7 @@ class CLI(ABC):
         loader.set_vault_secrets(vault_secrets)
 
         if self.USES_CONNECTION:
-            try:
-                _launch_ssh_agent()
-            except Exception as e:
-                raise AnsibleError('Failed to launch ssh agent.') from e
+            _agent_launch.launch_ssh_agent()
 
         # create the inventory, and filter it based on the subset specified (if any)
         inventory = InventoryManager(loader=loader, sources=options['inventory'], cache=(not options.get('flush_cache')))
