+++ lib/ansible/plugins/connection/ssh.py
@@ -447,7 +447,7 @@ from ansible.plugins.connection import ConnectionBase, BUFSIZE
 from ansible.plugins.shell.powershell import _replace_stderr_clixml
 from ansible.utils.display import Display
 from ansible.utils.path import unfrackpath, makedirs_safe
-from ansible.utils._ssh_agent import SshAgentClient, _key_data_into_crypto_objects
+from ansible._internal._ssh import _ssh_agent
 
 try:
     from cryptography.hazmat.primitives import serialization
@@ -766,12 +766,12 @@ class Connection(ConnectionBase):
         key_data = self.get_option('private_key')
         passphrase = self.get_option('private_key_passphrase')
 
-        private_key, public_key, fingerprint = _key_data_into_crypto_objects(
+        private_key, public_key, fingerprint = _ssh_agent.key_data_into_crypto_objects(
             to_bytes(key_data),
             to_bytes(passphrase) if passphrase else None,
         )
 
-        with SshAgentClient(auth_sock) as client:
+        with _ssh_agent.SshAgentClient(auth_sock) as client:
             if public_key not in client:
                 display.vvv(f'SSH: SSH_AGENT adding {fingerprint} to agent', host=self.host)
                 client.add(
